fix: reorganize localize test files to mirror source directory structure
Move test files from flattened naming to proper directory structure:
- tests/agents/localize-translate-documents.test.mjs → tests/agents/localize/translate-documents.test.mjs
- tests/agents/localize-translate-images.test.mjs → tests/agents/localize/translate-images.test.mjs

Update import paths accordingly.
